Rename package in MySQL module
