+++ extensions/cdc-mysql/src/main/java/com/hazelcast/jet/cdc/mysql/impl/MySqlSequenceExtractor.java
@@ -14,9 +14,10 @@
  * limitations under the License.
  */
 
-package com.hazelcast.jet.cdc.impl;
+package com.hazelcast.jet.cdc.mysql.impl;
 
 import com.hazelcast.internal.util.HashUtil;
+import com.hazelcast.jet.cdc.impl.SequenceExtractor;
 
 import java.util.Map;
 import java.util.Objects;
