+++ extensions/cdc-mysql/src/main/java/com/hazelcast/jet/cdc/mysql/package-info.java
@@ -18,4 +18,4 @@
  * Contains connectors for change data capture events from MySQL
  * databases.
  */
-package com.hazelcast.jet.cdc;
+package com.hazelcast.jet.cdc.mysql;
